requirements.txt
configparser
coverage
paramiko
tmdbsimple
rich-click
pytest-mock